move basic.rs
