+++ relation/src/circuit.rs
@@ -7,7 +7,7 @@
 //! Interfaces for plonk circuits
 
 use crate::errors::CircuitError;
-use ark_ff::{FftField, Field, PrimeField};
+use ark_ff::{FftField, Field};
 use ark_poly::univariate::DensePolynomial;
 
 /// An index to a gate in circuit.
@@ -161,60 +161,6 @@ pub trait Circuit<F: Field> {
     /// Constrain two variables to have the same value.
     /// Return error if the input variables are invalid.
     fn equal_gate(&mut self, a: Variable, b: Variable) -> Result<(), CircuitError>;
-
-    /// Constrain that `a` < `b`.
-    fn enforce_lt(&mut self, a: Variable, b: Variable) -> Result<(), CircuitError>
-    where
-        F: PrimeField;
-
-    /// Constrain that`a` <= `b`.
-    fn enforce_leq(&mut self, a: Variable, b: Variable) -> Result<(), CircuitError>
-    where
-        F: PrimeField,
-    {
-        let c = self.is_lt(b, a)?;
-        self.constant_gate(c.0, F::zero())
-    }
-
-    /// Constrain that `a` > `b`.
-    fn enforce_gt(&mut self, a: Variable, b: Variable) -> Result<(), CircuitError>
-    where
-        F: PrimeField,
-    {
-        self.enforce_lt(b, a)
-    }
-
-    /// Constrain that `a` >= `b`.
-    fn enforce_geq(&mut self, a: Variable, b: Variable) -> Result<(), CircuitError>
-    where
-        F: PrimeField,
-    {
-        let c = self.is_lt(a, b)?;
-        self.constant_gate(c.into(), F::zero())
-    }
-
-    /// Returns a `BoolVar` indicating whether `a` < `b`.
-    fn is_lt(&mut self, a: Variable, b: Variable) -> Result<BoolVar, CircuitError>
-    where
-        F: PrimeField;
-
-    /// Returns a `BoolVar` indicating whether `a` <= `b`.
-    fn is_leq(&mut self, a: Variable, b: Variable) -> Result<BoolVar, CircuitError>
-    where
-        F: PrimeField;
-
-    /// Returns a `BoolVar` indicating whether `a` > `b`.
-    fn is_gt(&mut self, a: Variable, b: Variable) -> Result<BoolVar, CircuitError>
-    where
-        F: PrimeField,
-    {
-        self.is_lt(b, a)
-    }
-
-    /// Returns a `BoolVar` indicating whether `a` >= `b`.
-    fn is_geq(&mut self, a: Variable, b: Variable) -> Result<BoolVar, CircuitError>
-    where
-        F: PrimeField;
 
     /// Pad the circuit with n dummy gates
     fn pad_gate(&mut self, n: usize);
