+++ relation/src/gadgets/mod.rs
@@ -6,8 +6,11 @@
 
 //! Gates and gadgets implementations
 
+pub mod basic;
 pub mod ecc;
 pub mod ultraplonk;
+
+pub use basic::PlonkCircuit;
 
 mod arithmetic;
 mod cmp;
